pyproject.toml
[project]
name = "gss-archetype-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "pandas>=3.0.2",
    "pyjwt>=2.12.1",
    "pyreadstat>=1.3.4",
    "scikit-learn>=1.8.0",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
